Add bunch of mysql structs

+++ src/MYSQL_BIND.cs
@@ -9,7 +9,7 @@ namespace MySqlSharp
     [StructLayout(LayoutKind.Sequential)]
     public unsafe struct MYSQL_BIND
     {
-        public ulong* length; /* output length pointer */
+        public uint* length; /* output length pointer */
         public bool* is_null;         /* Pointer to null indicator */
         public void* buffer;          /* buffer to get/put data */
         /* set this if you want to track data truncations happened during fetch */
@@ -19,9 +19,9 @@ namespace MySqlSharp
         public void* fetch_result;
         public void* skip_result;
         /* output buffer length, must be set when fetching str/binary */
-        public ulong buffer_length;
-        public ulong offset;              /* offset position for char/binary fetch */
-        public ulong length_value;        /* Used if length is 0 */
+        public uint buffer_length;
+        public uint offset;              /* offset position for char/binary fetch */
+        public uint length_value;        /* Used if length is 0 */
         public uint param_number;         /* For null count and error messages */
         public uint pack_length;          /* Internal length for packed data */
         public enum_field_types buffer_type; /* buffer type */
